added to roi_preprocess
